+++ config/app.php
@@ -3,7 +3,7 @@ declare(strict_types=1);
 return [
     'name' => 'TMS OS',
     'short_name' => 'TMS OS',
-    'build' => 'Platform V15.5.1',
+    'build' => 'Platform V15.5.2',
     'channel' => 'stable',
     'timezone' => 'Asia/Ho_Chi_Minh',
 ];